--- globals/utility.js
@@ -1,13 +1,0 @@
-import { getTarg } from '../utility/getTarg.js';
-import { hasPerm } from '../utility/hasPerm.js';
-import { hasRank } from '../utility/hasRank.js';
-import { takeRole } from '../utility/takeRole.js';
-
-const utility = {
-  'getTarg': getTarg,
-  'hasPerm': hasPerm,
-  'hasRank': hasRank,
-  'takeRole': takeRole,
-};
-
-export { utility };
